Unambiguous cast to tensor type long

+++ generate_with_bow_feedback.py
@@ -275,7 +275,7 @@ def build_bows_one_hot_vectors(bow_indices, tokenizer, device='cuda'):
     one_hot_bows_vectors = []
     for single_bow in bow_indices:
         single_bow = list(filter(lambda x: len(x) <= 1, single_bow))
-        single_bow = torch.tensor(single_bow).to(device)
+        single_bow = torch.tensor(single_bow, dtype=torch.long).to(device)
         num_words = single_bow.shape[0]
         one_hot_bow = torch.zeros(num_words, tokenizer.vocab_size).to(device)
         one_hot_bow.scatter_(1, single_bow, 1)
